relocated end2ast tests for where-function to correct test class
